moved bookings to services
